+++ libs/ngx-editorjs/CHANGELOG.md
@@ -5,12 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.1.1] - - 2019-05-10
 
 ### Added
 
-- Generated docs via [CompoDoc](https://github.com/compodoc/compodoc)
-- Internal updates
+- Generated docs via [CompoDoc](https://github.com/compodoc/compodoc) in npm package
+- Include CHANGELOG.md in npm package
 
 ## [1.1.0] - 2019-05-10